+++ src/noahtheduke/splint/rules/style/is_eq_order.clj
@@ -11,13 +11,13 @@
 (defn call-or-sym? [form]
   (or (symbol? form)
     (and (list? form)
-      (symbol? (first form)))))
+      (not (#{'quote 'syntax-quote 'splint/syntax-quote} (first form))))))
 
 (defrule style/is-eq-order
   "`clojure.test/is` expects `=`-based assertions to put the expected value first.
 
   This rule uses two checks on the `=` call to determine if it should issue a diagnostic:
-  * Is the first argument a symbol or a list with a symbol at the head? (A variable/local or a call.)
+  * Is the first argument a symbol or an unquoted list? (A variable/local or a call.)
   * Is the second argument a nil, boolean, char, number, keyword, or string?
 
   @examples